0.69 update [BTN + MIN Support]
Finally. . . the long awaited cosmic exploration gathering update!

This took awhile, and it's still not 100% complete. So bare with me here.
-> BTN + MIN has been re-added and spruced up to my standards.
-> Collectables are now fully supported, as well as reduction missions (this. was a pain in my ass)
-> Criticals aren't coded in yet, and dual class still needs a little bit of fine tuning before I ship those, so for the second those are unsupported/won't show up in your relic xp grinding
-> Fishing is... 80% along the way. I mainly wanted to get these 2 out first and foremost because I know that it's been asked. A lot. By everyone.
-> Added the ability to set a dismount range for heading to spots (if on a mount)
-> Auto-Gamba will now continue to gamba if it's detected, vs stopping if you had met the "stop if..." conditions

Thank you for your patience, and also thank you to everyone who's been helping nail out the gathering paths on the new moon while I've been busy coding. I appreciate every one of you for helping me out with this, it has saved a lot of time in the end <3 and I can't thank you enough.

This is going to stay in testing for a minute to make sure I didn't break shit. And I'm going to take a temp pause so I can -actually play some Silksong with friends instead of spending days coding between work xD-, I'll still be keeping an eye out for things/what all breaks with gathering but. As always if there's a problem do let me know o/